+++ src/main/java/com/verdantartifice/thaumicwonders/common/init/InitItems.java
@@ -16,6 +16,9 @@ import com.verdantartifice.thaumicwonders.common.items.entities.ItemFlyingCarpet
 import com.verdantartifice.thaumicwonders.common.items.misc.ItemDisjunctionCloth;
 import com.verdantartifice.thaumicwonders.common.items.misc.ItemStructureDiviner;
 import com.verdantartifice.thaumicwonders.common.items.misc.ItemTimewinder;
+import com.verdantartifice.thaumicwonders.common.items.plants.ItemCinderpearlSeed;
+import com.verdantartifice.thaumicwonders.common.items.plants.ItemShimmerleafSeed;
+import com.verdantartifice.thaumicwonders.common.items.plants.ItemVishroomSpore;
 import com.verdantartifice.thaumicwonders.common.items.tools.ItemBoneBow;
 import com.verdantartifice.thaumicwonders.common.items.tools.ItemPrimalDestroyer;
 
@@ -45,6 +48,9 @@ public class InitItems {
         registerItem(forgeRegistry, new ItemBoneBow());
         registerItem(forgeRegistry, new ItemPrimalArrow());
         registerItem(forgeRegistry, new ItemTW("primordial_grain"));
+        registerItem(forgeRegistry, new ItemShimmerleafSeed());
+        registerItem(forgeRegistry, new ItemCinderpearlSeed());
+        registerItem(forgeRegistry, new ItemVishroomSpore());
     }
     
     private static void registerItem(IForgeRegistry<Item> forgeRegistry, Item item) {