+++ changelog.md
@@ -3,6 +3,7 @@ v1.8.0
 * Added Flux Distiller, to clean out your capacitors
 * Added Primordial Accelerator, to smash your Primordial Pearls to bits, for science!
 * Added Primordial Accretion Chamber, to reconstitute your shattered pearls more efficiently
+* Added Mystic Gardening, to cultivate your magical plant life
 * Set rarity for several items
 * Fixed syncing bug when firing Primal Arrows from a Bone Bow
 * Clarified dyeing process for the Flying Carpet